requirements.txt
exif==1.6.0
pillow==11.0.0
plum-py==0.8.7
